+++ memote/suite/tests/test_growth.py
@@ -44,6 +44,14 @@ def test_growth_from_data_qualitative(model, experiment, threshold=0.95):
     metric but is a little fragile to not having any false negatives or false
     positives in the output.
 
+    Implementation:
+    Read and validate experimental config file and data tables. Constrain the
+    model with the parameters provided by a user's definition of the medium,
+    then compute a confusion matrix based on the predicted true, expected
+    true, predicted false and expected false growth.
+    The individual values of the confusion matrix are calculated as described
+    in https://en.wikipedia.org/wiki/Confusion_matrix
+
     """
     ann = test_growth_from_data_qualitative.annotation
     exp = pytest.memote.experimental.growth[experiment]